PCM-480 Flag Linked resource when no linked resource.
